+++ guest/rust/api_core/src/client/mod.rs
@@ -4,9 +4,6 @@ pub mod audio;
 /// **\[Client-only\]** Input retrieval and manipulation.
 pub mod input;
 
-/// **\[Client-only\]** Helper functions for the camera.
-pub mod camera;
-
 /// **\[Client-only\]** Manipulating the operating system's clipboard.
 pub mod clipboard;