+++ CHANGELOG.md
@@ -55,6 +55,8 @@ These PRs are not directly user-facing, but improve the development experience. 
 - **Docs**: The IDE documentation has been improved, including information on how to set up Emacs for Ambient development (thanks to [@kevzettler](https://github.com/kevzettler) in [#505](https://github.com/AmbientRun/Ambient/pull/505)).
 - **Assets**: `ambient assets import` can be used to import assets one by one. This will create or modify the `pipeline.toml` file for you.
 - **Packages**: A new fps_controller package to easily get started with a fps or tps camera
+- **Camera**: Added `camera::get_active` to get the active camera.
+- **Client**: When using the native client, you can now use `--window-x` and `--window-y` to specify the window position, as well as `--window-width` and `--window-height` to specify the window size.
 
 #### Examples
 
@@ -153,6 +155,7 @@ These PRs are not directly user-facing, but improve the development experience. 
 - **Logging**: The logging output levels have been tweaked to better communicate the state of the system at any given time.
 - **Debugger**: The debugger has been improved with a resizable sidebar, a scrollable view, and a component filter.
 - **Animation**: The animation graph is now executed on the server as well.
+- **CLI**: The `ambient new` command now takes several parameters to customize the resulting generation.
 
 ### Fixed